Created extension folder. Move date strideable extension into folder.
