+++ Sources/SunKit/Extensions.swift
@@ -19,28 +19,6 @@
 import Foundation
 
 
-//It consents us too loop between two dates for n as interval time
-extension Date: @retroactive Strideable {
-    public func distance(to other: Date) -> TimeInterval {
-        return other.timeIntervalSinceReferenceDate - self.timeIntervalSinceReferenceDate
-    }
-    
-    func toString(_ timeZone: TimeZone) -> String {
-        let df = DateFormatter()
-        df.timeZone = timeZone
-        let custom = DateFormatter.dateFormat(fromTemplate: "MMdd HH:mm",
-                                              options: 0,
-                                              locale: Locale(identifier: "en"))
-        df.dateFormat = custom
-        
-        return df.string(from: self)
-    }
-    
-    public func advanced(by n: TimeInterval) -> Date {
-        return self + n
-    }
-}
-
 extension Calendar {
     func numberOfDaysSinceStartOfTheYear(for date: Date) -> Int {
         let startOfTheYear: Date = startOfYear(date)
